Many Changes to Blog
Blog is almost complete. I put a text editor in the post section and am
able to upload photos. Changes to links of linkedIn and github on the
front. fixed the bootstrap on the blog page after realizing it was using
bootstrap 2.0 instead of 3.0 and ive added paging on the blog as well as
updated my photo on the front. The blog also had dotdotdot plugin added
and finally i updated my webconfig so it will post the correct database
online.