Fixed Slug and blog details
Fixed the slug URL so it is working properly as well as the blog details
page which displays my blog posts. Added a partial view for searching to
be displayed on my blog posts and blog details page. Moved the login and
register partial view from my navigation bar to my blog search partial
view