Many changes to Bug Boss
Added files for datatables, and added datatables. authorized all
controllers only to be accessed by people in roles, if not you are
redirected to the login page. added a projectViewModel and a
ticketViewModel. added all necessary user roles to config file as as
well ticket priorities in the config file. added time stamp to tickets.
added multiselect list to "users to add and remove". hid links from
unauthorized users in the layout. removed the updated text area because
it is not needed. fixed the webconfig so i can properly login through
the website.