Add Attachment, Notifications, Histories,
Able to add attachments to tickets, and I have the proper code in the
controller to send notifications upon the creation of a new ticket. All
I need fixed is the users and ticket notifications should work without a
problem. Ticket histories has the correct code in the controller as well
just need to touch up some things to make it work properly.